% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkBounded}
\alias{checkBounded}
\title{checkBounded}
\usage{
checkBounded(x, message, bound = c(0, 1))
}
\arguments{
\item{x}{x}

\item{message}{identifier for x}

\item{bound}{the boundaries, array of size two}
}
\description{
checks whether x is defined and lies within the specified bound
}
