% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkComparisonModel}
\alias{checkComparisonModel}
\title{checkComparisonModel}
\usage{
checkComparisonModel(comparison)
}
\arguments{
\item{comparison}{comparison}
}
\value{
valid type
}
\description{
checks whether comparison is one of 'restricted' or 'saturated' (or respective shortcuts)
}
