% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkDataGenerationTypes}
\alias{checkDataGenerationTypes}
\title{checkDataGenerationTypes}
\usage{
checkDataGenerationTypes(type)
}
\arguments{
\item{type}{type}
}
\value{
Returns cleaned data generation type
}
\description{
Checks whether data generation type is one of \code{'normal'}, \code{'IG'}, \code{'mnonr'}, or \code{'RK'}, stop otherwise.
}
