% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkMissingTypes}
\alias{checkMissingTypes}
\title{checkMissingTypes}
\usage{
checkMissingTypes(type)
}
\arguments{
\item{type}{type}
}
\value{
Returns cleaned data generation type
}
\description{
Checks whether missing generation type is one of \code{'mcar'}, \code{'mar'}, or \code{'nmar'}, stop otherwise.
}
