% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkNullEffect}
\alias{checkNullEffect}
\title{checkNullEffect}
\usage{
checkNullEffect(nullEffect, valid)
}
\arguments{
\item{nullEffect}{nullEffect}

\item{valid}{vector of valid effects}
}
\value{
nullEffect
}
\description{
checks whether nullEffect is one among valid effect
}
