% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPositive}
\alias{checkPositive}
\title{checkPositive}
\usage{
checkPositive(x, message = NULL)
}
\arguments{
\item{x}{x}

\item{message}{identifier for \code{x}}
}
\description{
Checks whether \code{x} is defined and a positive number, stop otherwise.
}
