% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPositiveDefinite}
\alias{checkPositiveDefinite}
\title{checkPositiveDefinite}
\usage{
checkPositiveDefinite(x, message = NULL, stop = TRUE)
}
\arguments{
\item{x}{x}

\item{message}{identifier for \code{x}}

\item{stop}{whether to stop or to throw a warning}
}
\description{
Checks whether \code{x} is positive definite, stop otherwise.
}
