% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPowerTypes}
\alias{checkPowerTypes}
\title{checkPowerTypes}
\usage{
checkPowerTypes(type)
}
\arguments{
\item{type}{type}
}
\value{
Returns cleaned type
}
\description{
Checks whether type is one of \code{'a-priori'}, \code{'post-hoc'}, or \code{'compromise'} (or respective shortcuts), stop otherwise.
}
