% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkSquare}
\alias{checkSquare}
\title{checkSquare}
\usage{
checkSquare(x, message = NULL)
}
\arguments{
\item{x}{x}

\item{message}{identifier for \code{x}}
}
\description{
Checks whether \code{x} is a square matrix, stop otherwise.
}
