% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkSymmetricSquare}
\alias{checkSymmetricSquare}
\title{checkSymmetricSquare}
\usage{
checkSymmetricSquare(x, message = NULL)
}
\arguments{
\item{x}{x}

\item{message}{identifier for x}
}
\description{
checks whether x is a symmetric square matrix
}
