% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{doSim}
\alias{doSim}
\title{doSim}
\usage{
doSim(
  r,
  simData,
  isMultigroup = FALSE,
  modelH0,
  modelH1,
  lavOptions,
  lavOptionsH1
)
}
\arguments{
\item{r}{replication id}

\item{simData}{list of datafiles}

\item{isMultigroup}{multigroup flag}

\item{modelH0}{\code{lavaan} model string defining the (incorrect) analysis model.}

\item{modelH1}{\code{lavaan} model string defining the comparison model. If omitted, the saturated model is the comparison model.}

\item{lavOptions}{a list of additional options passed to \code{lavaan}, e. g., \code{list(estimator = 'mlm')} to request robust ML estimation}

\item{lavOptionsH1}{lavoptions when fitting \code{modelH1}. If \code{NULL}, the same as \code{lavOptions}.}
}
\value{
list
}
\description{
Generates random data from population variance-covariance matrix and population means, either
from a multivariate normal distribution, or using one of various approaches to generate
non-normal data.
}
