% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genData.normal}
\alias{genData.normal}
\title{genData.normal}
\usage{
genData.normal(N = NULL, Sigma = NULL, nSets = 1)
}
\arguments{
\item{N}{sample size.}

\item{Sigma}{population covariance matrix.}

\item{nSets}{number of data sets to generate}
}
\value{
Returns the generated data
}
\description{
Generates multivariate normal random data conforming to a population variance-covariance matrix.
}
