% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genData}
\alias{genData}
\title{genData}
\usage{
genData(N = NULL, Sigma = NULL, mu = NULL, gIdx = NULL)
}
\arguments{
\item{N}{sample size.}

\item{Sigma}{population covariance matrix.}

\item{mu}{population means.}

\item{gIdx}{if not NULL, add gIdx as numeric group index as additional variable to generated data}
}
\value{
data
}
\description{
generate random data from population variance-covariance matrix and population means.
currently normal case only.
}
\examples{
\dontrun{
}
}
