% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genData.VM}
\alias{genData.VM}
\title{genData.VM}
\usage{
genData.VM(N = NULL, Sigma = NULL, nSets = 1, skewness = NULL, kurtosis = NULL)
}
\arguments{
\item{N}{sample size.}

\item{Sigma}{population covariance matrix.}

\item{nSets}{number of data sets to generate}

\item{skewness}{vector specifying skewness for each variable}

\item{kurtosis}{vector specifying excess kurtosis for each variable}
}
\value{
Returns the generated data
}
\description{
Generates random data conforming to a population variance-covariance matrix using
the third-order polynomial method  (Vale & Maurelli, 1983)
specifying third and fourth moments of the marginals.
}
\details{
This function is a slightly adapted copy of \code{lavaan}'s ValeMaurelli1983 implementation
that avoids computing the intermediate correlation for each data sets
and uses Sigma as input.

For details, see
Vale, C. & Maurelli, V. (1983). Simulating multivariate nonnormal distributions. \emph{Psychometrika, 48}, 465-471.
}
