% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getCFI.Sigma}
\alias{getCFI.Sigma}
\title{getCFI.Sigma}
\usage{
getCFI.Sigma(SigmaHat, S, muHat = NULL, mu = NULL)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}
}
\value{
CFI
}
\description{
calculates CFI given model-implied and observed covariance matrix.
}
\details{
cfi= (f_null - f_hyp) / f_null
}
