% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getChiSquare.F}
\alias{getChiSquare.F}
\title{getChiSquare.F}
\usage{
getChiSquare.F(Fmin, n, df)
}
\arguments{
\item{Fmin}{population minimum of the fit-function (can be a list for multiple group models).}

\item{n}{number of observations  (can be a list for multiple group models).}

\item{df}{model degrees of freedom}
}
\value{
Returns chi-square
}
\description{
Computes the (Wishart-) chi-square from the population minimum of the fit-function:
\code{chi-square = (N - 1) * F0 + df = ncp + df}. Note that F0 is the population minimum.
Using F_hat would give \code{chi-square = (N - 1) * F_hat}.
}
