% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{getDiscrepancyFunctionFromFittingFunction}
\alias{getDiscrepancyFunctionFromFittingFunction}
\title{getDiscrepancyFunctionFromFittingFunction}
\usage{
getDiscrepancyFunctionFromFittingFunction(fittingFunction)
}
\arguments{
\item{fittingFunction}{fittingFunction}
}
\value{
discrepancy function
}
\description{
get proper discrepancy function (to measure F0) from fitting function (to obtain SigmaHat)
}
