% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compromise.R
\name{getErrorDiff}
\alias{getErrorDiff}
\title{getErrorDiff}
\usage{
getErrorDiff(critChiSquare, df, ncp, log.abratio)
}
\arguments{
\item{critChiSquare}{evaluated chi-squared value}

\item{df}{the model degrees of freedom}

\item{ncp}{the non-centrality parameter}

\item{log.abratio}{log(alpha/beta)}
}
\value{
squared difference between alpha and beta on a log scale
}
\description{
Determine the squared log-difference between alpha and beta error given a certain chi-square value from central chi-square(df) and a non-central chi-square(df, ncp) distribution.
}
