% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.AGFI}
\alias{getF.AGFI}
\title{getF.AGFI}
\usage{
getF.AGFI(AGFI, df, p)
}
\arguments{
\item{AGFI}{AGFI}

\item{df}{model degrees of freedom}

\item{p}{number of observed variables}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function from AGFI
}
\details{
F_min = rmsea^2 * df
}
\examples{
\dontrun{
fmin <- getF.AGFI(AGFI = .95, df = 100, p = 15)
fmin
}
}
