% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.gammaHat}
\alias{getF.gammaHat}
\title{getF.gammaHat}
\usage{
getF.gammaHat(gammaHat, p)
}
\arguments{
\item{gammaHat}{gammaHat}

\item{p}{number of observed variables}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function from gammaHat
}
\examples{
\dontrun{
F_min <- getF.gammaHat(gammaHat = .90, p=15)
F_min
}
}
