% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.GFI}
\alias{getF.GFI}
\title{getF.GFI}
\usage{
getF.GFI(GFI, p)
}
\arguments{
\item{GFI}{GFI}

\item{p}{number of observed variables}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function from AGFI
}
