% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.Mc}
\alias{getF.Mc}
\title{getF.Mc}
\usage{
getF.Mc(Mc)
}
\arguments{
\item{Mc}{Mc}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function from Mc
}
\examples{
\dontrun{
F_min <- getF.Mc(Mc = .90)
F_min
}
}
