% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF}
\alias{getF}
\title{getF}
\usage{
getF(effect, effect.measure, df = NULL, p = NULL, SigmaHat = NULL,
  Sigma = NULL)
}
\description{
calculates minimum of the ML-fit-function from known fit indices
}
