% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.RMSEA}
\alias{getF.RMSEA}
\title{getF.RMSEA}
\usage{
getF.RMSEA(RMSEA, df)
}
\arguments{
\item{RMSEA}{RMSEA}

\item{df}{model degrees of freedom}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function from RMSEA
}
\details{
F_min = rmsea^2 * df
}
\examples{
\dontrun{
fmin <- getF.RMSEA(RMSEA = .05, df = 100)
fmin
}
}
