% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{getFittingFunctionFromEstimator}
\alias{getFittingFunctionFromEstimator}
\title{getFittingFunctionFromEstimator}
\usage{
getFittingFunctionFromEstimator(lavOptions)
}
\arguments{
\item{lavOptions}{lavOptions}
}
\value{
fitting function
}
\description{
get proper fitting function (to obtain sigmaHat) for chosen estimator
}
