% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{getFormattedResults}
\alias{getFormattedResults}
\title{getFormattedResults}
\usage{
getFormattedResults(type, result, digits = 6)
}
\arguments{
\item{type}{type of power analysis}

\item{result}{result object (list)}

\item{digits}{number of significant digits}
}
\value{
data.frame
}
\description{
Return data.frame containing formatted results.
}
