% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{getFormattedSimulationResults}
\alias{getFormattedSimulationResults}
\title{getFormattedResults}
\usage{
getFormattedSimulationResults(object, digits = 2)
}
\arguments{
\item{object}{result object (list)}

\item{digits}{number of significant digits}

\item{type}{type of power analysis}
}
\value{
data.frame
}
\description{
Return data.frame containing formatted results.
}
