% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getGammaHat.F}
\alias{getGammaHat.F}
\title{getGammaHat.F}
\usage{
getGammaHat.F(Fmin, p)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}

\item{p}{number of observed variables}
}
\value{
GammaHat
}
\description{
calculates GammaHat from minimum of the ML-fit-function
}
\examples{
\dontrun{
GammaHat <- getGammaHat.F(Fmin = .05, p = 15)
GammaHat
}
}
