% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getGFI.F}
\alias{getGFI.F}
\title{getGFI.F}
\usage{
getGFI.F(Fmin, p)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}

\item{p}{number of observed variables}
}
\value{
Returns GFI
}
\description{
Computes GFI from the minimum of the ML-fit-function.
}
