% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getIndices.F}
\alias{getIndices.F}
\title{getIndices.F}
\usage{
getIndices.F(
  fmin,
  df,
  p = NULL,
  SigmaHat = NULL,
  Sigma = NULL,
  muHat = NULL,
  mu = NULL,
  N = NULL
)
}
\arguments{
\item{fmin}{minimum of the ML-fit-function}

\item{df}{model degrees of freedom}

\item{p}{number of observed variables}

\item{SigmaHat}{model implied covariance matrix}

\item{Sigma}{population covariance matrix}

\item{muHat}{model implied means}

\item{mu}{population means}

\item{N}{list of sample weights}
}
\value{
list of indices
}
\description{
Computes known indices from the minimum of the ML-fit-function.
}
