% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getKSdistance}
\alias{getKSdistance}
\title{getKSdistance}
\usage{
getKSdistance(chi, df, ncp = 0)
}
\arguments{
\item{chi}{empirical distribution}

\item{df}{df of reference distribution}

\item{ncp}{ncp of reference distribution}
}
\value{
average absolute distance
}
\description{
computes average absulute KS-distance between empirical and asympotic chi-square reference distribution.
}
