% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getLavOptions}
\alias{getLavOptions}
\title{getLavOptions}
\usage{
getLavOptions(lavOptions = NULL, isCovarianceMatrix = TRUE, nGroups = 1)
}
\arguments{
\item{lavOptions}{additional options to be added to (or overwriting) the defaults}

\item{isCovarianceMatrix}{if \code{TRUE}, also adds \code{sample.nobs = 1000} and \code{sample.cov.rescale = FALSE} to \code{lavoptions}}

\item{nGroups}{the number of groups, 1 by default}
}
\value{
a list of \code{lavaan} defaults
}
\description{
returns \code{lavaan} options including defaults as set in \code{sem()} as a list to be passed to \code{lavaan()}
}
