% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getMc.F}
\alias{getMc.F}
\title{getMc.F}
\usage{
getMc.F(Fmin)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}
}
\value{
Mc
}
\description{
calculates Mc from minimum of the ML-fit-function
}
