% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getNCP}
\alias{getNCP}
\title{getNCP}
\usage{
getNCP(Fmin, n)
}
\arguments{
\item{Fmin}{population minimum of the fit-function (can be a list for multiple group models).}

\item{n}{number of observations (can be a list for multiple group models).}
}
\value{
Returns the implied NCP.
}
\description{
Computes the non-centrality parameter from the population minimum of the fit-function
(dividing by N - 1 following the Wishart likelihood): \code{ncp = (N - 1) * F0}.
}
