% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getRMSEA.F}
\alias{getRMSEA.F}
\title{getRMSEA.F}
\usage{
getRMSEA.F(Fmin, df, nGroups = 1)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}

\item{df}{model degrees of freedom}

\item{nGroups}{the number of groups}
}
\value{
RMSEA
}
\description{
calculates RMSEA from minimum of the ML-fit-function
}
\details{
F_min = rmsea^2 * df
}
