% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getSRMR.Sigma}
\alias{getSRMR.Sigma}
\title{getSRMR.Sigma}
\usage{
getSRMR.Sigma(SigmaHat, S, muHat = NULL, mu = NULL)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}
}
\value{
Returns SRMR
}
\description{
Computes SRMR given the model-implied and the observed (or population) covariance matrix,
using the Hu & Bentler approach to standardization.
}
