% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getWLSv}
\alias{getWLSv}
\title{getWLSv}
\usage{
getWLSv(S, mu = NULL, diag = FALSE)
}
\arguments{
\item{S}{observed (or population) covariance matrix}

\item{mu}{observed (or population) mean vector}

\item{diag}{weight matrix for DWLS}
}
\value{
Returns V
}
\description{
Computes the WLS weight matrix as the asymptotic covariance matrix of the sample statistics
}
