% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{makeRestrictionsLavFriendly}
\alias{makeRestrictionsLavFriendly}
\title{makeRestrictionsLavFriendly}
\usage{
makeRestrictionsLavFriendly(model)
}
\arguments{
\item{model}{\code{lavaan} model string}
}
\value{
model with  \code{lavaan}-friendly constrains
}
\description{
This function transforms a \code{lavaan} model string into a model string that works reliably
when both equality constrains and value constrains are imposed on the same parameters.
\code{lavaan} cannot reliably handle this case, e. g., "a == b \n a == 0" will not always work.
The solution is to drop the equality constraint and rather apply
the value constraint on each equality constrained parameter, e. g. "a == 0 \n b == 0" will work.
}
