% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{orderLavCov}
\alias{orderLavCov}
\title{orderLavCov}
\usage{
orderLavCov(lavCov = NULL)
}
\arguments{
\item{lavCov}{model implied covariance matrix}
}
\value{
cov in correct order
}
\description{
returns \code{lavaan} implied covariance matrix in correct order.
}
