% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{orderLavMu}
\alias{orderLavMu}
\title{orderLavMu}
\usage{
orderLavMu(lavMu = NULL)
}
\arguments{
\item{lavMu}{model implied means}
}
\value{
mu in correct order
}
\description{
returns \code{lavaan} implied means in correct order.
}
