% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{orderLavResults}
\alias{orderLavResults}
\title{orderLavResults}
\usage{
orderLavResults(lavCov = NULL, lavMu = NULL)
}
\arguments{
\item{lavCov}{model implied covariance matrix}

\item{lavMu}{model implied means}
}
\value{
either cov or mu in correct order
}
\description{
returns \code{lavaan} implied covariance matrix or mean vector in correct order.
}
