% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{powerPrepare}
\alias{powerPrepare}
\title{powerPrepare}
\usage{
powerPrepare(
  type = NULL,
  effect = NULL,
  effect.measure = NULL,
  alpha = NULL,
  beta = NULL,
  power = NULL,
  abratio = NULL,
  N = NULL,
  df = NULL,
  p = NULL,
  SigmaHat = NULL,
  Sigma = NULL,
  muHat = NULL,
  mu = NULL,
  simulatedPower = FALSE,
  modelH0 = NULL,
  modelH1 = NULL,
  nReplications = NULL,
  minConvergenceRate = NULL,
  lavOptions = NULL
)
}
\arguments{
\item{type}{type of power analysis}

\item{effect}{effect size specifying the discrepancy between H0 and H1 (a list for multiple group models; a vector of length 2 for effect-size differences)}

\item{effect.measure}{type of effect, one of "F0", "RMSEA", "Mc", "GFI", AGFI"}

\item{alpha}{alpha error}

\item{beta}{beta error; set either beta or power}

\item{power}{power (1-beta); set either beta or power}

\item{abratio}{the ratio of alpha to beta}

\item{N}{the number of observations (a list for multiple group models)}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for effect.measure = "GFI" and "AGFI"}

\item{SigmaHat}{model implied covariance matrix (a list for multiple group models). Use in conjunction with Sigma to define effect and effect.measure.}

\item{Sigma}{population covariance matrix (a list for multiple group models). Use in conjunction with SigmaHat to define effect and effect.measure.}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}

\item{simulatedPower}{whether to perform a simulated (TRUE) (rather than analytical, FALSE) power analysis.}

\item{modelH0}{for simulated power: lavaan model string defining the (incorrect) analysis model.}

\item{modelH1}{for simulated power: lavaan model string defining the comparison model. If omitted, the saturated model is the comparison model.}
}
\value{
list
}
\description{
do some preparations common to all types of power analyses
}
