% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHoc.R
\name{semPower.postHoc}
\alias{semPower.postHoc}
\title{semPower.postHoc}
\usage{
semPower.postHoc(
  effect = NULL,
  effect.measure = NULL,
  alpha,
  N,
  df,
  p = NULL,
  SigmaHat = NULL,
  Sigma = NULL,
  muHat = NULL,
  mu = NULL,
  ...
)
}
\arguments{
\item{effect}{effect size specifying the discrepancy between H0 and H1 (a list for multiple group models; a vector of length 2 for effect-size differences)}

\item{effect.measure}{type of effect, one of "F0", "RMSEA", "Mc", "GFI", AGFI"}

\item{alpha}{alpha error}

\item{N}{the number of observations (a list for multiple group models)}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for effect.measure = "GFI" and "AGFI"}

\item{SigmaHat}{model implied covariance matrix (a list for multiple group models). Use in conjunction with Sigma to define effect and effect.measure.}

\item{Sigma}{population covariance matrix (a list for multiple group models). Use in conjunction with SigmaHat to define effect and effect.measure.}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}
}
\value{
list
}
\description{
Determine power (1-beta) given alpha, df, and effect
}
\examples{
\dontrun{
power <- semPower.postHoc(effect = .05, effect.measure = "RMSEA", alpha = .05, N = 250, df = 200)
summary(power)
power <- semPower.postHoc(effect = list(.02, .01), effect.measure = "F0", 
                          alpha = .05, N = list(250, 350), df = 200)
summary(power)
power <- semPower.postHoc(N = 1000, df = 5, alpha = .05,  
                          SigmaHat = diag(4), Sigma = cov(matrix(rnorm(4*1000),  ncol=4)))
summary(power)
}
}
