% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{semPower.powerLav}
\alias{semPower.powerLav}
\title{semPower.powerLav}
\usage{
semPower.powerLav(
  type,
  modelPop = NULL,
  modelH0 = NULL,
  modelH1 = NULL,
  fitH1model = TRUE,
  Sigma = NULL,
  mu = NULL,
  ...
)
}
\arguments{
\item{type}{type of power analysis, one of 'a-priori', 'post-hoc', 'compromise'.}

\item{modelPop}{lavaan model string defining the true model.}

\item{modelH0}{lavaan model string defining the (incorrect) analysis model.}

\item{modelH1}{lavaan model string defining the comparison model. If omitted, the saturated model is the comparison model.}

\item{fitH1model}{whether to fit the H1 model. If FALSE, the H1 model is assumed to show the same fit as the saturated model, and the delta df are set to 1.}

\item{Sigma}{population covariance matrix (if modelPop is not set).}

\item{mu}{population means (if modelPop is not set).}

\item{...}{other parameters related to the specific type of power analysis requested}
}
\value{
a list containing the results of the power analysis, the population covariance matrix Sigma, the H0 implied matrix SigmaHat, as well as various lavaan model strings
}
\description{
Perform power analysis on population and model-implied Sigmas as defined through lavaan model strings
This requires the lavaan package.
}
\examples{
\dontrun{
## a priori power analysis for the null hypothesis that the correlation between 
## two cfa factors with a true correlation of .2 differs from zero
# define population model 
mPop <- '
  f1 =~ .5*x1 + .6*x2 + .4*x3
  f2 =~ .7*x4 + .8*x5 + .3*x6
  x1 ~~ .75*x1
  x2 ~~ .64*x2
  x3 ~~ .84*x3
  x4 ~~ .51*x4
  x5 ~~ .36*x5
  x6 ~~ .91*x6
  f1 ~~ 1*f1
  f2 ~~ 1*f2
  f1 ~~ .2*f2
'
# define analysis model (restricting the factor correlation to zero) 
mAna <- '
  f1 =~ x1 + x2 + x3
  f2 =~ x4 + x5 + x6
  f1 ~~ 0*f2
'
# do a priori power analsis
lavpower.ap <- semPower.powerLav(type = 'a-priori', 
                                 modelPop = mPop, modelH0 = mAna,
                                 alpha = .05, beta = .05)
summary(lavpower.ap$power)

}
}
