% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{semPower.powerMI}
\alias{semPower.powerMI}
\title{semPower.powerMI}
\usage{
semPower.powerMI(type, comparison = NULL, nullEffect = NULL, ...)
}
\arguments{
\item{type}{type of power analysis, one of 'a-priori', 'post-hoc', 'compromise'}

\item{comparison}{comparison model, either 'saturated', or one of 'configural', 'metric', 'scalar', or a vector of restrictions in lavaan format (with 'none' for no restrictions). See details.}

\item{nullEffect}{defines the hypothesis of interest. One of 'metric', 'scalar', 'residual', or a vector of restrictions in lavaan format (with 'none' for no restrictions).  See details.}

\item{...}{other parameters specifying the factor model (see [semPower.genSigma()]) and the type of power analysis}
}
\value{
a list containing the results of the power analysis, the population covariance matrix Sigma and mean vector mu, the H0 implied matrix SigmaHat and mean vector muHat, as well as various lavaan model strings (modelH0, and modelH1)
}
\description{
Convenience function for performing power analysis for multigroup measurement invariance models.
This requires the lavaan package.
}
\details{
The models defined in the comparison and the nullEffect arguments can be specified as follows:
\itemize{
\item `configural`: no invariance constraints. Shows the same fit as the saturated model, so only the delta df differ. 
\item `metric`: all loadings are restricted to equality. 
\item `scalar`: all loadings and (item-)intercepts are restricted to equality. 
\item `residual`: all loadings, (item-)intercepts, and (item-)residuals are restricted to equality.
}
For greater flexibility, the models can also be defined using lavaan style group.equal restrictions as a vector: 
\itemize{
\item `'none'`: no invariance constraints. Shows the same fit as the saturated model, so only the delta df differ. 
\item `c('loadings')`: all loadings are restricted to equality. 
\item `c('loadings', 'intercepts')`: all loadings and (item-)intercepts are restricted to equality. 
\item `c('loadings', 'intercepts', 'residuals')`: all loadings, (item-)intercepts, and (item-)residuals are restricted to equality.
\item `c('loadings', 'residuals')`: all loadings and (item-)residuals are restricted to equality.
\item `c('loadings', 'residuals', 'lv.variances)`: all loadings, (item-)residuals, and latent variances are restricted to equality.
\item `c('loadings', 'intercepts', 'means)`: all loadings, (item-)intercepts, and latent means are restricted to equality.
}
}
\examples{
\dontrun{
}
}
\seealso{
[semPower.genSigma()]
}
