% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{semPower.powerPlot.byEffect}
\alias{semPower.powerPlot.byEffect}
\title{semPower.powerPlot.byEffect}
\usage{
semPower.powerPlot.byEffect(
  effect.measure = NULL,
  alpha,
  N,
  df,
  p = NULL,
  effect.min = NULL,
  effect.max = NULL,
  steps = 50,
  linewidth = 1
)
}
\arguments{
\item{effect.measure}{type of effect, one of \code{"F0"}, \code{"RMSEA"}, \code{"Mc"}, \code{"GFI"}, \code{"AGFI"}}

\item{alpha}{alpha error}

\item{N}{the number of observations}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for \code{effect.measure = "GFI"} and \code{effect.measure = "AGFI"}}

\item{effect.min}{minimum effect}

\item{effect.max}{maximum effect}

\item{steps}{number of steps}

\item{linewidth}{linewidth}
}
\value{
powerplot
}
\description{
Shows a plot showing power as function of the effect for a given N and alpha.
}
\examples{
\dontrun{
semPower.powerPlot.byEffect(effect.measure = "RMSEA", alpha = .05, 
                            N = 500, effect.min = .01, effect.max = .15, df = 200)
}
}
