% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{semPower.powerPlot}
\alias{semPower.powerPlot}
\title{semPower.powerPlot}
\usage{
semPower.powerPlot(semPowerRes, ...)
}
\arguments{
\item{semPowerRes}{results of a semPower analysis}

\item{...}{other parameters passed to \code{\link[=semPower.powerPlot.byN]{semPower.powerPlot.byN()}}}
}
\value{
powerplot
}
\description{
Shows a plot showing power as function of N given the output of a power analysis.
}
\examples{
\dontrun{
# perform a power analysis
powerCFA <- semPower.powerCFA(type = 'post-hoc', alpha = .05, N = 300,
                              Phi = .15, nIndicator = c(5, 4), loadM = c(.5, .6))
# show plot
semPower.powerPlot(powerCFA)
}
}
