% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compromise.R
\name{summary.semPower.compromise}
\alias{summary.semPower.compromise}
\title{summary.sempower.compromise}
\usage{
\method{summary}{semPower.compromise}(object, ...)
}
\arguments{
\item{object}{result object from semPower.compromise}

\item{...}{other}
}
\description{
provide summary of compromise post-hoc power analyses
}
