% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validateInput}
\alias{validateInput}
\title{validateInput}
\usage{
validateInput(power.type = NULL, effect = NULL, effect.measure = NULL,
  alpha = NULL, beta = NULL, power = NULL, abratio = NULL, N = NULL,
  df = NULL, p = NULL, SigmaHat = NULL, Sigma = NULL)
}
\arguments{
\item{power.type}{type of power analyses, one of "a-priori", post-hoc", "compromise"}

\item{effect}{effect size specifying the discrepancy between H0 and H1}

\item{effect.measure}{type of effect, one of "F0, "RMSEA", "Mc", "GFI", AGFI"}

\item{alpha}{alpha error}

\item{beta}{beta error}

\item{power}{power (1-beta)}

\item{abratio}{ratio alpha/beta}

\item{N}{the number of observations}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for effect.measure = "GFI" and "AGFI"}

\item{SigmaHat}{model implied covariance matrix}

\item{Sigma}{population covariance matrix}
}
\description{
Validates input for power calcuation function
}
