% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkComparisonModel}
\alias{checkComparisonModel}
\title{checkComparisonModel}
\usage{
checkComparisonModel(comparison)
}
\arguments{
\item{comparison}{comparison}
}
\value{
Returns cleaned comparison
}
\description{
Checks whether comparison is one of \code{'restricted'} or \code{'saturated'} (or respective shortcuts), stop otherwise.
}
