% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkEllipsis}
\alias{checkEllipsis}
\title{checkEllipsis}
\usage{
checkEllipsis(...)
}
\arguments{
\item{...}{the parameters to search}
}
\value{
whether loadings and power are found
}
\description{
checks whether ... contains arguments related to loadings and to power.
}
