% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPositive}
\alias{checkPositive}
\title{checkPositive}
\usage{
checkPositive(x, message)
}
\arguments{
\item{x}{x}

\item{message}{identifier for x}
}
\description{
checks whether x is defined and a positive number, stop otherwise
}
