% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPositiveDefinite}
\alias{checkPositiveDefinite}
\title{checkPositiveDefinite}
\usage{
checkPositiveDefinite(x, message = NULL)
}
\arguments{
\item{x}{x}

\item{message}{identifier for x}
}
\description{
checks whether x is positive definite
}
