% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{checkPowerTypes}
\alias{checkPowerTypes}
\title{checkPowerTypes}
\usage{
checkPowerTypes(type)
}
\arguments{
\item{type}{type}
}
\value{
valid type
}
\description{
checks whether type is one of 'a-priori', 'post-hoc', or 'compromise'.
}
