% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{genData}
\alias{genData}
\title{genData}
\usage{
genData(N = NULL, Sigma = NULL, mu = NULL, gIdx = NULL)
}
\arguments{
\item{N}{sample size.}

\item{Sigma}{population covariance matrix.}

\item{mu}{population means.}

\item{gIdx}{if not \code{NULL}, add gIdx as numeric group index as additional variable to generated data}
}
\value{
Returns the generated data
}
\description{
Generates random data from population variance-covariance matrix and population means.
Currently only the multivariate normal case.
}
\examples{
\dontrun{
gen <- semPower.genSigma(Phi = .2, loadings = list(rep(.5, 3), rep(.7, 3)))
data <- genData(N = 500, Sigma = gen$Sigma) 
}
}
