% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getAGFI.F}
\alias{getAGFI.F}
\title{getAGFI.F}
\usage{
getAGFI.F(Fmin, df, p)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}

\item{df}{model degrees of freedom}

\item{p}{number of observed variables}
}
\value{
AGFI
}
\description{
calculates AGFI from minimum of the ML-fit-function
}
