% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPriori.R
\name{getBetadiff}
\alias{getBetadiff}
\title{getBetadiff}
\usage{
getBetadiff(cN, critChi, logBetaTarget, fmin, df, weights = NULL)
}
\arguments{
\item{cN}{current N}

\item{critChi}{critical chi-square associated with chosen alpha error}

\item{logBetaTarget}{log(desired beta)}

\item{fmin}{minimum of the ML fit function}

\item{df}{the model degrees of freedom}

\item{weights}{sample weights for multiple group models}
}
\value{
squared difference requested and achieved beta on a log scale
}
\description{
get squared difference between requested and achieved beta on a logscale
}
