% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getCFI.Sigma.mgroups}
\alias{getCFI.Sigma.mgroups}
\title{getCFI.Sigma.mgroups}
\usage{
getCFI.Sigma.mgroups(SigmaHat, S, muHat = NULL, mu = NULL, N)
}
\arguments{
\item{SigmaHat}{a list of model implied covariance matrix}

\item{S}{a list of observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}

\item{N}{a list of group weights}
}
\value{
CFI
}
\description{
calculates CFI given model-implied and observed covariance matrix for multiple group models.
}
\details{
cfi= (f_null - f_hyp) / f_null
}
