% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getCFI.Sigma}
\alias{getCFI.Sigma}
\title{getCFI.Sigma}
\usage{
getCFI.Sigma(SigmaHat, S)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}
}
\value{
CFI
}
\description{
calculates CFI given model-implied and observed covariance matrix.
}
\details{
cfi= (f_null - f_hyp) / f_null
}
\examples{
\dontrun{
CFI <- getCFI.Sigma(SigmaHat = diag(4), S = cov(matrix(rnorm(4*100),  ncol=4)))
CFI
}
#' @export
}
