% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getCFI.Sigma}
\alias{getCFI.Sigma}
\title{getCFI.Sigma}
\usage{
getCFI.Sigma(SigmaHat, S, muHat = NULL, mu = NULL, fittingFunction = "ML")
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}

\item{fittingFunction}{whether to use \code{ML} or \code{WLS}}
}
\value{
Returns CFI
}
\description{
Computes CFI given the model-implied and the observed (or population) covariance matrix:
\code{CFI = (F_null - F_hyp) / F_null}.
}
