% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getChiSquare.F}
\alias{getChiSquare.F}
\title{getChiSquare.F}
\usage{
getChiSquare.F(Fmin, n, df)
}
\arguments{
\item{Fmin}{population minimum of the fit-function}

\item{n}{number of observations}

\item{df}{model degrees of freedom}
}
\value{
NCP
}
\description{
calculates chis-square from the population minimum of the fit-function
}
\details{
chi = (n-1)*F +  df = ncp + df

note that F is the population minimum; using F_hat would give chi = (n-1)*F_hat
}
\examples{
\dontrun{
NCP <- getChiSquare.F(Fmin = .05, n = 1000, df = 100)
NCP
}
}
