% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getChiSquare.NCP}
\alias{getChiSquare.NCP}
\title{getChiSquare.NCP}
\usage{
getChiSquare.NCP(NCP, df)
}
\arguments{
\item{NCP}{non-centrality parameter}

\item{df}{model degrees of freedom}
}
\value{
Returns chi-square
}
\description{
Computes chi-square from the non-centrality parameter: \code{chi-square = ncp + df}.
}
