% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getF.GFI}
\alias{getF.GFI}
\title{getF.GFI}
\usage{
getF.GFI(GFI, p)
}
\arguments{
\item{GFI}{GFI}

\item{p}{number of observed variables}
}
\value{
Returns Fmin
}
\description{
Computes the minimum of the ML-fit-function from GFI.
}
