% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getF}
\alias{getF}
\title{getF}
\usage{
getF(
  effect,
  effect.measure,
  df = NULL,
  p = NULL,
  SigmaHat = NULL,
  Sigma = NULL,
  muHat = NULL,
  mu = NULL
)
}
\arguments{
\item{effect}{magnitude of effect}

\item{effect.measure}{measure of effect, one of \code{'fmin'}, \code{'rmsea'}, \code{'agfi'}, \code{'gfi'}, \code{'mc'}}

\item{df}{model degrees of freedom}

\item{p}{number of observed variables}

\item{SigmaHat}{model implied covariance matrix}

\item{Sigma}{observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}
}
\value{
Returns Fmin
}
\description{
Computes the minimum of the ML-fit-function from known fit indices.
}
