% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getF.RMSEA}
\alias{getF.RMSEA}
\title{getF.RMSEA}
\usage{
getF.RMSEA(RMSEA, df)
}
\arguments{
\item{RMSEA}{RMSEA}

\item{df}{model degrees of freedom}
}
\value{
Returns Fmin
}
\description{
Computes the minimum of the ML-fit-function from RMSEA:
\code{F_min = rmsea^2 * df}.
}
