% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getF.Sigma}
\alias{getF.Sigma}
\title{getF.Sigma}
\usage{
getF.Sigma(SigmaHat, S, muHat = NULL, mu = NULL)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function given model-implied and observed covariance matrix.
}
\details{
F_min = tr(S %*% SigmaHat^-1) - p + ln(det(SigmaHat)) - ln(det(S))
with meanstructure add (mu - muHat)' SigmaHat^-1 (mu - muHat)
}
