% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getF.Sigma}
\alias{getF.Sigma}
\title{getF.Sigma}
\usage{
getF.Sigma(SigmaHat, S)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}
}
\value{
Fmin
}
\description{
calculates minimum of the ML-fit-function given model-implied and observed covariance matrix.
}
\details{
F_min = tr(S %*% SigmaHat^-1) - p + ln(det(SigmaHat)) - ln(det(S))
}
\examples{
\dontrun{
fmin <- getF.Sigma(SigmaHat = diag(4), S = cov(matrix(rnorm(4*100),  ncol=4)))
fmin
}
}
