% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getIndices.F}
\alias{getIndices.F}
\title{getIndices.F}
\usage{
getIndices.F(fmin, df, p = NULL, SigmaHat = NULL, Sigma = NULL)
}
\arguments{
\item{df}{model degrees of freedom}

\item{p}{number of observed variables}

\item{Fmin}{minimum of the ML-fit-function}
}
\value{
list of indices
}
\description{
calculates known indices from minimum of the ML-fit-function
}
