% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getLavOptions}
\alias{getLavOptions}
\title{getLavOptions}
\usage{
getLavOptions(lavOptions = NULL, isCovarianceMatrix = TRUE, nGroups = 1)
}
\arguments{
\item{lavOptions}{additional options to be added to (or overwriting) the defaults}

\item{isCovarianceMatrix}{if TRUE, also adds sample.nobs = 1000 and sample.cov.rescale = FALSE to lavoptions}

\item{nGroupes}{the number of groups, 1 by default}
}
\value{
a list of lav defaults
}
\description{
returns lavaan options including defaults as set in sem() as a list to be passed to lavaan()
}
\examples{
\dontrun{
}
}
