% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getNCP}
\alias{getNCP}
\title{getNCP}
\usage{
getNCP(Fmin, n)
}
\arguments{
\item{Fmin}{population minimum of the fit-function}

\item{n}{number of observations}
}
\value{
NCP
}
\description{
calculates non-centrality parameter from the population minimum of the fit-function
}
\details{
ncp = (n-1) * F
}
