% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getPhi.B}
\alias{getPhi.B}
\title{getPhi.B}
\usage{
getPhi.B(B)
}
\arguments{
\item{B}{matrix of regression coefficients (all-y notation). Must only contain non-zero lower-triangular elements, so the first row only includes zeros.}
}
\value{
the implied correlation matrix
}
\description{
Computes implied correlations from Beta matrix (using all-y notation), disallowing recursive paths.
}
\examples{
\dontrun{
B <- matrix(c(
c(.00, .00, .00),
c(.10, .00, .00),
c(.20, .30, .00)
), byrow = TRUE, ncol = 3)
Phi <- getPhi.B(B)
}
}
