% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getRMSEA.F}
\alias{getRMSEA.F}
\title{getRMSEA.F}
\usage{
getRMSEA.F(Fmin, df, nGroups = 1)
}
\arguments{
\item{Fmin}{minimum of the ML-fit-function}

\item{df}{model degrees of freedom}

\item{nGroups}{the number of groups}
}
\value{
Returns RMSEA
}
\description{
Computes RMSEA from the minimum of the ML-fit-function
\code{F_min = rmsea^2 * df}.
}
