% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSizes.R
\name{getSRMR.Sigma.mgroups}
\alias{getSRMR.Sigma.mgroups}
\title{getSRMR.Sigma.mgroups}
\usage{
getSRMR.Sigma.mgroups(SigmaHat, S, muHat = NULL, mu = NULL, N)
}
\arguments{
\item{SigmaHat}{a list of model implied covariance matrices}

\item{S}{a list of observed (or population) covariance matrices}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}

\item{N}{a list of group weights}
}
\value{
Returns SRMR
}
\description{
Computes SRMR given the model-implied and the observed (or population) covariance matrix for multiple group models
using the Hu & Bentler approach to standardization and the MPlus approach to multiple group sampling weights
(weight squared sums of residuals).
}
