% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getSRMR.Sigma.mgroups}
\alias{getSRMR.Sigma.mgroups}
\title{getSRMR.Sigma.mgroups}
\usage{
getSRMR.Sigma.mgroups(SigmaHat, S, N)
}
\arguments{
\item{SigmaHat}{a list of model implied covariance matrices}

\item{S}{a list of observed (or population) covariance matrices}

\item{N}{a list of group weights}
}
\value{
SRMR
}
\description{
calculates SRMR given model-implied and observed covariance matrix for multiple group models
}
