% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getSRMR.Sigma}
\alias{getSRMR.Sigma}
\title{getSRMR.Sigma}
\usage{
getSRMR.Sigma(SigmaHat, S)
}
\arguments{
\item{SigmaHat}{model implied covariance matrix}

\item{S}{observed (or population) covariance matrix}
}
\value{
SRMR
}
\description{
calculates SRMR given model-implied and observed covariance matrix.
}
\examples{
\dontrun{
SRMR <- getSRMR.Sigma(SigmaHat = diag(4), S = cov(matrix(rnorm(4*100),  ncol=4)))
SRMR
}
}
