% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPriori.R
\name{semPower.aPriori}
\alias{semPower.aPriori}
\title{semPower.aPriori}
\usage{
semPower.aPriori(
  effect = NULL,
  effect.measure = NULL,
  alpha,
  beta = NULL,
  power = NULL,
  N = NULL,
  df = NULL,
  p = NULL,
  SigmaHat = NULL,
  Sigma = NULL,
  muHat = NULL,
  mu = NULL,
  simulatedPower = FALSE,
  modelH0 = NULL,
  modelH1 = NULL,
  nReplications = 250,
  minConvergenceRate = 0.5,
  lavOptions = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{effect}{effect size specifying the discrepancy between H0 and H1 (a list for multiple group models; a vector of length 2 for effect-size differences)}

\item{effect.measure}{type of effect, one of "F0", "RMSEA", "Mc", "GFI", AGFI"}

\item{alpha}{alpha error}

\item{beta}{beta error; set either beta or power}

\item{power}{power (1-beta); set either beta or power}

\item{N}{a list of sample weights for multiple group power analyses, e.g. list(1,2) to make the second group twice as large as the first one}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for effect.measure = "GFI" and "AGFI"}

\item{SigmaHat}{model implied covariance matrix (a list for multiple group models). Use in conjunction with Sigma to define effect and effect.measure.}

\item{Sigma}{population covariance matrix (a list for multiple group models). Use in conjunction with SigmaHat to define effect and effect.measure.}

\item{muHat}{model implied mean vector}

\item{mu}{observed (or population) mean vector}

\item{simulatedPower}{whether to perform a simulated (TRUE) (rather than analytical, FALSE) power analysis.}

\item{modelH0}{for simulated power: lavaan model string defining the (incorrect) analysis model.}

\item{modelH1}{for simulated power: lavaan model string defining the comparison model. If omitted, the saturated model is the comparison model.}

\item{nReplications}{for simulated power: number of random samples drawn.}

\item{minConvergenceRate}{for simulated power: the minimum convergence rate required}

\item{lavOptions}{for simulated power: a list of additional options passed to lavaan, e.g., list(estimator = 'mlm') to request robust ML estimation}

\item{seed}{for simulated power}
}
\value{
list
}
\description{
Determine required sample size given alpha, beta/power, df, and effect
}
\examples{
\dontrun{
power <- semPower.aPriori(effect = .05, effect.measure = "RMSEA", alpha = .05, beta = .05, df = 200)
summary(power)
power <- semPower.aPriori(effect = .15, effect.measure = "F0", alpha = .05, power = .80, df = 100)
summary(power)
power <- semPower.aPriori(effect = list(.05, .10), effect.measure = "F0", alpha = .05, 
                          power = .80, N = list(1, 1), df = 100)
summary(power)
power <- semPower.aPriori(alpha = .01, beta = .05, df = 5, 
                          SigmaHat = diag(4), Sigma = cov(matrix(rnorm(4*1000),  ncol=4)))
summary(power)
}
}
