% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{semPower.getDf}
\alias{semPower.getDf}
\title{semPower.getDf}
\usage{
semPower.getDf(lavModel, nGroups = NULL, group.equal = NULL)
}
\arguments{
\item{lavModel}{the lavaan model string}

\item{nGroups}{for multigroup models: the number of groups}

\item{group.equal}{for multigroup models: vector of strings defining the type(s) of cross-group equality constraints following the lavaan conventions (loadings, intercepts, means, residuals, residual.covariances, lv.variances, lv.covariances, regressions)}
}
\value{
the df of the model
}
\description{
Determines the degrees of freedom of a given model provided as lavaan model string. 
This requires the lavaan package.
}
\examples{
\dontrun{
lavModel <- '
f1 =~ x1 + x2 + x3 + x4
f2 =~ x5 + x6 + x7 + x8
f3 =~ y1 + y2 + y3
f3 ~ f1 + f2
'
semPower.getDf(lavModel)

# multigroup version
semPower.getDf(lavModel, nGroups = 3)  
semPower.getDf(lavModel, nGroups = 3, group.equal = c('loadings'))
semPower.getDf(lavModel, nGroups = 3, group.equal = c('loadings', 'intercepts'))
}
}
