% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{semPower.getDf}
\alias{semPower.getDf}
\title{semPower.getDf}
\usage{
semPower.getDf(lavModel)
}
\arguments{
\item{lavModel}{the lavaan model string}
}
\value{
df
}
\description{
Convenience function to determine the degrees of freedom of a given model provided as lavaan model string. 
This requires the lavaan package.
}
\examples{
\dontrun{
lavModel <- '
f1 =~ x1 + x2 + x3 + x4
f2 =~ x5 + x6 + x7 + x8
f3 =~ y1 + y2 + y3
f3 ~ f1 + f2
'
semPower.getDf(lavModel)
}
}
