% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenienceFunctions.R
\name{semPower.powerMediation}
\alias{semPower.powerMediation}
\title{semPower.powerMediation}
\usage{
semPower.powerMediation(
  type,
  comparison = "restricted",
  bYX = NULL,
  bMX = NULL,
  bYM = NULL,
  Beta = NULL,
  indirect = NULL,
  nullEffect = "ind = 0",
  nullWhichGroups = NULL,
  ...
)
}
\arguments{
\item{type}{type of power analysis, one of \code{'a-priori'}, \code{'post-hoc'}, \code{'compromise'}.}

\item{comparison}{comparison model, one of \code{'saturated'} or \code{'restricted'} (the default). This determines the df for power analyses. \code{'Saturated'} provides power to reject the model when compared to the saturated model, so the df equal the one of the hypothesized model. \code{'Restricted'} provides power to reject the hypothesized model when compared to an otherwise identical model that just omits the restrictions defined in \code{nullEffect}, so the df equal the number of restrictions.}

\item{bYX}{the standardized slope (direct effect) for X -> Y. A list for multiple group models. Can be \code{NULL} if \code{Beta} is set.}

\item{bMX}{the standardized slope for X -> M. A list for multiple group models. Can be \code{NULL} if \code{Beta} is set.}

\item{bYM}{the standardized slope for M -> Y. A list for multiple group models. Can be \code{NULL} if \code{Beta} is set.}

\item{Beta}{can be used instead of \code{bYX}, \code{bMX}, and \code{bYM}: matrix of regression weights connecting the latent factors (all-Y notation). Exogenous variables must be in the first row(s), so the upper triangular of Beta must be zero. A list for multiple group models.}

\item{indirect}{\code{NULL} unless \code{Beta} is set. Otherwise a list of vectors of size 2 indicating the elements of \code{Beta} that define the indirect effect of interest, e.g. \code{list(c(2, 1), c(3, 2))}. See details.}

\item{nullEffect}{defines the hypothesis of interest, must be one of \code{'ind = 0'} (the default) to test whether the indirect effect is zero or \code{'indA = indB'} to test for the equality of indirect effects across groups. See details.}

\item{nullWhichGroups}{for \code{nullEffect = 'indA = indB'}, vector indicating the groups for which equality constrains should be applied, e.g. \code{c(1, 3)} to constrain the relevant parameters of the first and the third group. If \code{NULL}, all groups are constrained to equality.}

\item{...}{other parameters related to the specific type of power analysis requested, see \code{\link[=semPower.aPriori]{semPower.aPriori()}}, \code{\link[=semPower.postHoc]{semPower.postHoc()}}, and \code{\link[=semPower.compromise]{semPower.compromise()}}, and specifying the factor model (see \code{\link[=semPower.genSigma]{semPower.genSigma()}}).  Note that in case of a simple mediation, the order of factors is X, M, Y.}
}
\value{
A list containing the following components is returned:
\item{\code{power}}{the results of the power analysis. Use the \code{summary} method to obtain formatted results.}
\item{\code{Sigma}}{the population covariance matrix. A list for multiple group models.}
\item{\code{mu}}{the population mean vector or \code{NULL} when no meanstructure is involved. A list for multiple group models.}
\item{\code{SigmaHat}}{the H0 model implied covariance matrix. A list for multiple group models.}
\item{\code{muHat}}{the H0 model implied mean vector or \code{NULL} when no meanstructure is involved. A list for multiple group models.}
\item{\code{modelH0}}{\code{lavaan} H0 model string.}
\item{\code{modelH1}}{\code{lavaan} H1 model string or \code{NULL} when the comparison refers to the saturated model.}
}
\description{
Convenience function for performing power analysis concerning indirect effect(s) in a mediation model.
This requires the lavaan package.
}
\details{
Notes on implementation:
\itemize{
\item For models without latent variables, \code{nullEffect = 'ind = 0'} and \code{nullEffect = 'indA = indB'} constrain the indirect effect to zero and to equality, respectively.
\item For models with latent variables and \code{nullEffect = 'ind = 0'}, power is approximated by constraining the smallest slope contained in the indirect effect to zero.
\item For models with latent variables multiple groups (i.e., \code{nullEffect = 'indA = indB'}), there is currently no way to determine power,
because implementing equality constrains on the indirect effects leads to non-convergence and
the approach to constrain a single or all slopes to equality across groups misrepresents the actual hypothesis of interest.
}
}
\examples{
\dontrun{
# simple case of X -> M -> Y mediation
# X -- .30 --> M -- .40 --> Y 
# X --------- .25 --------> Y 
# providing the number of indicators by factor (X, M, Y) each loading by the same magnitude on its designed factor.
medPower <- semPower.powerMediation(type = 'a-priori', 
                                    bYX = .25, bMX = .3, bYM = .4,
                                    nIndicator = c(3, 5, 4), 
                                    loadM = c(.5, .6, .7),                                     
                                    alpha = .05, beta = .05)
summary(medPower$power)
                                    
# same mediation model as above, but assuming single loading of 1 for all factors 
# (=> mediation model with manifest variables)
medPower <- semPower.powerMediation(type = 'a-priori', 
                                    bYX = .25, bMX = .3, bYM = .4,
                                    nIndicator = c(1, 1, 1), loadM = 1,
                                    alpha = .05, beta = .05)

# same latent mediation model as above, but specifying loadings through Lambda 
Lambda <- matrix(c(
                 #  X,   M,   Y
                c(0.5, 0.0, 0.0),    
                c(0.4, 0.0, 0.0),
                c(0.3, 0.0, 0.0),
                c(0.0, 0.7, 0.0),
                c(0.0, 0.8, 0.0),
                c(0.0, 0.5, 0.0),
                c(0.0, 0.0, 0.5),
                c(0.0, 0.0, 0.4),
                c(0.0, 0.0, 0.6)
               ), byrow = TRUE, ncol = 3)
medPower <- semPower.powerMediation(type = 'a-priori', 
                                    bYX = .25, bMX = .3, bYM = .4,
                                    Lambda = Lambda,                                     
                                    alpha = .05, beta = .05)

# same mediation model as above, but specifying Beta 
B <- matrix(c(
              c(.00, .00, .00),
              c(.30, .00, .00),
              c(.25, .40, .00)
              ), byrow = TRUE, ncol = 3)
medPower <- semPower.powerMediation(type = 'a-priori', 
                                    Beta = B, indirect = list(c(2,1), c(3,2)),
                                    Lambda = Lambda,
                                    alpha = .05, beta = .05)

# Beta for a more complex mediation hypothesis
# of the form X -> M1 -> M2 -> Y 
# and using a reduced loading matrix
B <- matrix(c(
              c(.00, .00, .00, .00),       # X
              c(.20, .00, .00, .00),       # M1
              c(.00, .30, .00, .00),       # M2
              c(.00, .00, .40, .00)        # Y
              ), byrow = TRUE, ncol = 4)
# only define primary loadings by factor
loadings <- list(
               c(0.4, 0.5, 0.8),           # X
               c(0.7, 0.6, 0.5, 0.8),      # M1
               c(0.5, 0.6, 0.3, 0.4, 0.6), # M2
               c(0.6, 0.7, 0.8)            # Y
               )

medPower <- semPower.powerMediation(type = 'a-priori', 
                                    Beta = B, indirect = list(c(2,1), c(3,2), c(4,3)),
                                    loadings = loadings,
                                    alpha = .05, beta = .05)
 
# multigroup example
medPower <- semPower.powerMediation(type = 'a-priori', 
                                    nullEffect = 'indA = indB',
                                    bYX = list(.25, .25), bMX = list(.3, .3), bYM = list(.4, .5),
                                    Lambda = diag(3), N = list(1, 1),
                                    alpha = .05, beta = .05)
}
}
\seealso{
\code{\link[=semPower.genSigma]{semPower.genSigma()}} \code{\link[=semPower.aPriori]{semPower.aPriori()}} \code{\link[=semPower.postHoc]{semPower.postHoc()}} \code{\link[=semPower.compromise]{semPower.compromise()}}
}
