% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{semPower.powerPlot.byEffect}
\alias{semPower.powerPlot.byEffect}
\title{sempower.powerPlot.byEffect}
\usage{
semPower.powerPlot.byEffect(
  effect.measure = NULL,
  alpha,
  N,
  df,
  p = NULL,
  effect.min = NULL,
  effect.max = NULL,
  steps = 50,
  linewidth = 1
)
}
\arguments{
\item{effect.measure}{type of effect, one of "F0", "RMSEA", "Mc", "GFI", AGFI"}

\item{alpha}{alpha error}

\item{N}{the number of observations}

\item{df}{the model degrees of freedom}

\item{p}{the number of observed variables, required for effect.measure = "GFI" and "AGFI"}

\item{effect.min}{minimum effect}

\item{effect.max}{maximum effect}

\item{steps}{number of steps}

\item{linewidth}{linewidth}
}
\value{
powerplot
}
\description{
show a plot showing power as function of N for a given effect and alpha
}
\examples{
\dontrun{
semPower.powerPlot.byEffect(effect.measure = "RMSEA", alpha = .05, 
                            N = 500, effect.min = .01, effect.max = .15, df = 200)
}
}
