% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{semPower.showPlot}
\alias{semPower.showPlot}
\title{semPower.showPlot}
\usage{
semPower.showPlot(chiCrit, ncp, df, linewidth = 1, showLabels = TRUE)
}
\arguments{
\item{chiCrit}{critical chi-square, e. g. \code{qchisq(alpha, df, ncp = 0, lower.tail = FALSE)}}

\item{ncp}{non-centrality parameter under H1}

\item{df}{degrees of freedom}

\item{linewidth}{linewidth}

\item{showLabels}{whether to add labels}
}
\description{
Shows a plot showing central and non-central chi-square distribution.
}
