% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  modelH0 = NULL,
  modelH1 = NULL,
  Sigma = NULL,
  mu = NULL,
  N = NULL,
  alpha = NULL,
  nReplications = 250,
  minConvergenceRate = 0.5,
  lavOptions = NULL,
  lavOptionsH1 = lavOptions,
  returnFmin = TRUE
)
}
\arguments{
\item{modelH0}{lavaan model string defining the (incorrect) analysis model.}

\item{modelH1}{lavaan model string defining the comparison model. If omitted, the saturated model is the comparison model.}

\item{Sigma}{population covariance matrix.}

\item{mu}{population means.}

\item{N}{sample size}

\item{alpha}{alpha error probability}

\item{nReplications}{number of random samples drawn.}

\item{minConvergenceRate}{the minimum convergence rate required}

\item{lavOptions}{a list of additional options passed to lavaan, e.g., list(estimator = 'mlm') to request robust ML estimation}

\item{lavOptionsH1}{lavoptions for H1model. If NULL, the same as lavOptions.}

\item{returnFmin}{whether return the median unbiased Fmin over replications (i.e, fmin_0 = fmin_hat - df/N)}
}
\value{
empirical power: sum(p < alpha) / nReplications
}
\description{
estimate empirical power using a simulation approach
}
\examples{
\dontrun{
}
}
