% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aPriori.R
\name{summary.semPower.aPriori}
\alias{summary.semPower.aPriori}
\title{summary.semPower.aPriori}
\usage{
\method{summary}{semPower.aPriori}(object, ...)
}
\arguments{
\item{result}{result object from semPower.aPriori}
}
\description{
provide summary of a-priori power analyses
}
