% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHoc.R
\name{summary.semPower.postHoc}
\alias{summary.semPower.postHoc}
\title{semPower.postHoc.summary}
\usage{
\method{summary}{semPower.postHoc}(object, ...)
}
\arguments{
\item{object}{result object from semPower.posthoc}

\item{...}{other}
}
\description{
provide summary of post-hoc power analyses
}
